requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
Brotli==1.1.0
certifi==2023.7.22
cffi==1.15.1
colorama==0.4.6
cryptography==41.0.5
cssselect2==0.7.0
decorator==5.1.1
dirtree==0.3.0.dev0
distlib==0.3.7
dj-database-url==2.1.0
Django==4.2.7
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.6.0
executing==2.0.0
filelock==3.12.4
fonttools==4.42.1
gunicorn==21.2.0
html5lib==1.1
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
packaging==23.1
pango==0.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pipenv==2023.9.8
platformdirs==3.10.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.8
pure-eval==0.2.2
pycparser==2.21
pydyf==0.8.0
Pygments==2.16.1
PyJWT==2.0.1
pyphen==0.14.0
python-decouple==3.8
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
stack-data==0.6.3
tinycss2==1.2.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
validate-email==1.3
virtualenv==20.24.5
wcwidth==0.2.8
weasyprint==60.0
webencodings==0.5.1
whitenoise==6.5.0
xlwt==1.3.0
zopfli==0.2.3
